+++ src/containers/SearchPage/SearchPage.duck.js
@@ -3,7 +3,13 @@ import unionWith from 'lodash/unionWith';
 import config from '../../config';
 import { storableError } from '../../util/errors';
 import { convertUnitToSubUnit, unitDivisor } from '../../util/currency';
-import { parseDateFromISO8601, getExclusiveEndDate } from '../../util/dates';
+import {
+  parseDateFromISO8601,
+  getExclusiveEndDate,
+  addTime,
+  subtractTime,
+  daysBetween,
+} from '../../util/dates';
 import { createImageVariantConfig } from '../../util/sdkLoader';
 import { isOriginInUse, isStockInUse } from '../../util/search';
 import { parse } from '../../util/urlHelpers';
@@ -34,8 +40,6 @@ const initialState = {
   searchInProgress: false,
   searchListingsError: null,
   currentPageResultIds: [],
-  searchMapListingIds: [],
-  searchMapListingsError: null,
 };
 
 const resultIds = data => data.data.map(l => l.id);
@@ -62,27 +66,6 @@ const listingPageReducer = (state = initialState, action = {}) => {
       // eslint-disable-next-line no-console
       console.error(payload);
       return { ...state, searchInProgress: false, searchListingsError: payload };
-
-    case SEARCH_MAP_LISTINGS_REQUEST:
-      return {
-        ...state,
-        searchMapListingsError: null,
-      };
-    case SEARCH_MAP_LISTINGS_SUCCESS: {
-      const searchMapListingIds = unionWith(
-        state.searchMapListingIds,
-        resultIds(payload.data),
-        (id1, id2) => id1.uuid === id2.uuid
-      );
-      return {
-        ...state,
-        searchMapListingIds,
-      };
-    }
-    case SEARCH_MAP_LISTINGS_ERROR:
-      // eslint-disable-next-line no-console
-      console.error(payload);
-      return { ...state, searchMapListingsError: payload };
 
     case SEARCH_MAP_SET_ACTIVE_LISTING:
       return {
@@ -114,19 +97,6 @@ export const searchListingsError = e => ({
   payload: e,
 });
 
-export const searchMapListingsRequest = () => ({ type: SEARCH_MAP_LISTINGS_REQUEST });
-
-export const searchMapListingsSuccess = response => ({
-  type: SEARCH_MAP_LISTINGS_SUCCESS,
-  payload: { data: response.data },
-});
-
-export const searchMapListingsError = e => ({
-  type: SEARCH_MAP_LISTINGS_ERROR,
-  error: true,
-  payload: e,
-});
-
 export const searchListings = searchParams => (dispatch, getState, sdk) => {
   dispatch(searchListingsRequest(searchParams));
 
@@ -142,23 +112,52 @@ export const searchListings = searchParams => (dispatch, getState, sdk) => {
   };
 
   const datesSearchParams = datesParam => {
+    const searchTZ = 'Etc/UTC';
+    const datesFilter = config.custom.defaultFilters.find(f => f.key === 'dates');
     const values = datesParam ? datesParam.split(',') : [];
-    const hasValues = datesParam && values.length === 2;
-    const startDate = hasValues ? values[0] : null;
-    const isNightlyBooking = config.lineItemUnitType === 'line-item/night';
+    const hasValues = datesFilter && datesParam && values.length === 2;
+    const { mode, entireRangeAvailable } = datesFilter || {};
+    const isNightlyMode = mode === 'night';
+
+    // SearchPage need to use a single time zone but listings can have different time zones
+    // We need to expand/prolong the time window (start & end) to cover other time zones too.
+    //
+    // NOTE: you might want to consider changing UI so that
+    //   1) location is always asked first before date range
+    //   2) use some 3rd party service to convert location to time zone (IANA tz name)
+    //   3) Make exact dates filtering against that specific time zone
+    //   This setup would be better for dates filter,
+    //   but it enforces a UX where location is always asked first and therefore configurability
+    const getProlongedStart = date => subtractTime(date, 14, 'hours', searchTZ);
+    const getProlongedEnd = date => addTime(date, 12, 'hours', searchTZ);
+
+    const startDate = hasValues ? parseDateFromISO8601(values[0], searchTZ) : null;
+    const endRaw = hasValues ? parseDateFromISO8601(values[1], searchTZ) : null;
     const endDate =
-      hasValues && isNightlyBooking
-        ? values[1]
+      hasValues && isNightlyMode
+        ? endRaw
         : hasValues
-        ? getExclusiveEndDate(values[1], 'Etc/UTC')
+        ? getExclusiveEndDate(endRaw, searchTZ)
         : null;
 
+    const dayCount = entireRangeAvailable ? daysBetween(startDate, endDate) : 1;
+    const day = 1440;
+    const hour = 60;
+    // When entire range is required to be available, we count minutes of included date range,
+    // but there's a need to subtract one hour due to possibility of daylight saving time.
+    // If partial range is needed, then we just make sure that the shortest time unit supported
+    // is available within the range.
+    // You might want to customize this to match with your time units (e.g. day: 1440 - 60)
+    const minDuration = entireRangeAvailable ? dayCount * day - hour : hour;
     return hasValues
       ? {
-          start: parseDateFromISO8601(startDate, 'Etc/UTC'),
-          end: parseDateFromISO8601(endDate, 'Etc/UTC'),
-          // Availability can be full or partial. Default value is full.
-          availability: 'full',
+          start: getProlongedStart(startDate),
+          end: getProlongedEnd(endDate),
+          // Availability can be time-full or time-partial.
+          // However, due to prolonged time window, we need to use time-partial.
+          availability: 'time-partial',
+          // minDuration uses minutes
+          minDuration,
         }
       : {};
   };
@@ -194,28 +193,6 @@ export const setActiveListing = listingId => ({
   payload: listingId,
 });
 
-export const searchMapListings = searchParams => (dispatch, getState, sdk) => {
-  dispatch(searchMapListingsRequest(searchParams));
-
-  const { perPage, ...rest } = searchParams;
-  const params = {
-    ...rest,
-    per_page: perPage,
-  };
-
-  return sdk.listings
-    .query(params)
-    .then(response => {
-      dispatch(addMarketplaceEntities(response));
-      dispatch(searchMapListingsSuccess(response));
-      return response;
-    })
-    .catch(e => {
-      dispatch(searchMapListingsError(storableError(e)));
-      throw e;
-    });
-};
-
 export const loadData = (params, search) => {
   const queryParams = parse(search, {
     latlng: ['origin'],
@@ -232,7 +209,7 @@ export const loadData = (params, search) => {
   const aspectRatio = aspectHeight / aspectWidth;
 
   return searchListings({
-    // TODO ...minStockMaybe,
+    ...minStockMaybe,
     ...rest,
     ...originMaybe,
     page,